+++ NEWS.md
@@ -3,7 +3,9 @@ v 0.1.1
 - Improving unit testing and code coverage
 - Improving documentation
 - Solving minor bug in date conversion and in which functions
-- New feature: unFactor to unfactor columns, when reading wasn't performed in expected way.
+- New features: 
+	* unFactor to unfactor columns, when reading wasn't performed in expected way.
+	* sameShape to make ure that train and test set have exactly the same shape.
 
 
 v 0.1: release on CRAN July 2017